backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "pyrefly>=0.47.0",
    "pytest>=9.0.2",
    "python-multipart>=0.0.21",
    "sqlalchemy[asyncio]>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]
